Feedback from pull request #1245
