+++ IPython/core/magic.py
@@ -57,6 +57,7 @@ from IPython.core.pylabtools import mpl_runner
 from IPython.testing.skipdoctest import skip_doctest
 from IPython.utils import py3compat
 from IPython.utils import openpy
+from IPython.utils.encoding import getdefaultencoding
 from IPython.utils.io import file_read, nlprint
 from IPython.utils.module_paths import find_mod
 from IPython.utils.path import get_py_filename, unquote_filename
@@ -949,7 +950,7 @@ Currently the magic system has the following functions:\n"""
                 try:
                     vstr = str(var)
                 except UnicodeEncodeError:
-                    vstr = unicode(var).encode(py3compat.getdefaultencoding(),
+                    vstr = unicode(var).encode(getdefaultencoding(),
                                                'backslashreplace')
                 vstr = vstr.replace('\n','\\n')
                 if len(vstr) < 50:
