+++ IPython/utils/_process_posix.py
@@ -26,6 +26,7 @@ from .autoattr import auto_attr
 from ._process_common import getoutput, arg_split
 from IPython.utils import text
 from IPython.utils import py3compat
+from IPython.utils.encoding import getdefaultencoding
 
 #-----------------------------------------------------------------------------
 # Function definitions
@@ -128,7 +129,7 @@ class ProcessHandler(object):
         int : child's exitstatus
         """
         # Get likely encoding for the output.
-        enc = py3compat.getdefaultencoding()
+        enc = getdefaultencoding()
         
         # Patterns to match on the output, for pexpect.  We read input and
         # allow either a short timeout or EOF
